1_2_jumpmap_demo
made a 1-2, jump map demo with free unity assets. the scarecrow is the spawn point of slimes. dummy for the goblins, and added thorns. the blue dot is the jangpoong level ups, and the treasure chest is the locations of potions. after the actual potion tiles come out i'll replace them with actual potions. this is just a demo version of the map.